+++ CHANGES.md
@@ -1,5 +1,8 @@
+##2.1.2
+- [#43](https://github.com/serilog/serilog-sinks-splunk/pull/43) - Extend sink & static configuration to allow for custom JSON formatter.
+
 ##2.1.1
-- [#38](https://github.com/serilog/serilog-sinks-splunk/issues/38)
+- [#38](https://github.com/serilog/serilog-sinks-splunk/issues/38) - Fix for HttpEventlogCollector and sourceType
 - Clean up of sample app using examples of host, sourcetype, source override
 
 ##2.1.0